requirements.txt
python-telegram-bot>=20.3
pypdf>=3.17.0
reportlab>=4.0.0
Flask==3.0.0
python-dotenv==1.0.0